update test coverage

+++ app/models/employee.py
@@ -1,7 +1,7 @@
 from app.utils.database import db
 
 class employee(db.Model):
-    id = db.Column(db.BigInteger, primary_key= True)
+    id = db.Column(db.Integer, primary_key= True)
     name = db.Column(db.String(50), nullable = False)
     age =  db.Column(db.Integer, nullable = False)
     specialization = db.Column(db.String(100), nullable=False)
